Mejoras en la gestión de reservas y servicios
Se ha añadido el uso de `SGHR.Persistence.Interfaces.Repositories.Habitaciones` en las pruebas de `ServicioApplicationServiceTest.cs`.
El método `ObtenerTodasReservasAsync` en `IReservaApplicationService.cs` ahora incluye un parámetro opcional `bool incluirRelaciones`.
Se ha implementado `ObtenerServiciosActivosAsync` en `IServicioApplicationService.cs`, eliminando métodos obsoletos relacionados con precios por categoría.
Se han actualizado los métodos en `ReservaApplicationService.cs` y `ServicioApplicationService.cs` para aceptar el nuevo parámetro y mejorar la inyección de dependencias.
Se han corregido parámetros en `Reservas.cs` y se ha añadido la inyección de `IServicioCategoriaApplicationService` en `DependencyInjection.cs`.
Se han implementado pruebas unitarias en `ReservaRepositoryTest.cs` y `ServicioRepositoryTest.cs`.
El método `ObtenerTodasAsync` en `IReservaRepository.cs` ahora acepta un parámetro para incluir relaciones.
Se han eliminado métodos de activación de servicios en `ServicioRepository.cs`, y se han realizado mejoras en los controladores `ReservasController.cs` y `ServiciosController.cs`.
Se ha añadido el controlador `ServicioCategoriaController.cs` para gestionar precios de servicios por categoría, junto con su correspondiente lógica en `ServicioCategoriaApplicationService.cs`.

+++ SGHR.Domain/Entities/Reservas/Reservas.cs
@@ -22,7 +22,7 @@ namespace SGHR.Domain.Entities.Reservas
         public CategoriaHabitacion? CategoriaHabitacion { get; private set; }
 
         protected Reserva() { }
-        public Reserva(int clienteId, int idCategoriaHabitacion, DateTime fechaEntrada, DateTime fechaSalida, int numeroHuespedes )
+        public Reserva(int clienteId, int idCategoriaHabitacion, DateTime fechaEntrada, DateTime fechaSalida, int numeroHuespedes)
         {
             if (clienteId <= 0)
             {
@@ -73,7 +73,7 @@ namespace SGHR.Domain.Entities.Reservas
             FechaCancelacion = DateTime.Now;
         }
 
-        public void ActualizarDetalles(int clienteId, int idCategoriaHabitacion, DateTime fechaEntrada, DateTime fechaSalida, sbyte numeroHuespedes)
+        public void ActualizarDetalles(int clienteId, int idCategoriaHabitacion, DateTime fechaEntrada, DateTime fechaSalida, int numeroHuespedes)
         {
             if (clienteId <= 0)
             {
